requirements.txt
wandb
scipy
webrtcvad
tensorflow==2.8.0
tensorflow_addons
numpy
absl-py
transformers
tensorflow_model_optimization
django
djangorestframework
google-cloud-secret-manager==2.12.4
psutil
gunicorn